requirements.txt
numpy
torch>=1.5
tensorboardX
easydict
pyyaml
tqdm
tfrecord
